requirements.txt
streamlit==1.32.0
face-recognition==1.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
pandas==2.2.1
dlib==19.24.2
